Fix taking Jacobian with rvalue expressions
When running PrepareExpr on an expression containing a leaf by value (denoted
`Expr<Leaf&&>`), *don't* copy the Leaf&& to PreparedType.  Instead, the
"prepared" expression contains an lvalue reference to the leaf in the original
expression (`Expr<Leaf>`). This change allows applying isSame() between the
prepared and original expression.

Drawback: the prepared expression can no longer exist outside of the original
expression's lifetime. Currently this is not a problem, but something to be
careful about in the future.