Fix clang 5 warning
In this case we wanted `Framed` to refer to the class template. Clang 3 wrongly treated it as a type-name. When changed to `Framed::template Framed`, Clang 5
warned that that actually means a constructor. An alternative solution is to qualify the name (`::wave::Framed`).

Here, avoid the whole issue by reusing the existing `Tag` alias in `traits`.