Deploy docs from Travis
Build Sphinx docs and deploy to github pages.

To show multiple versions of docs (in the future) use sphinxcontrib-versioning
(my own branch with bugfixes).

Use Travis build stages and clean up the other jobs a bit.  Give them names and
move sudo:required to only the job that needs it.

Add docs/README.md which will show up on github for the docs directory and the
gh-pages branch.