Bump version after release
Bump version to next expected release, and add -dev suffix to docs.

Todo: use some system that automatically determines the version, so it's not
necessary to bump the version before and after a release. This problem is not
trivial, because I would want to cover all the following cases

- Git checkout of pre-release commit
- Git checkout of release tag
- Non-git archive of pre-release commit (as offered by github for every commit)
- Non-git archive of release

For each case, documentation must be generated with the correct version number.
While tools exist to determine the version number from git, I am not satisfied
with the non-git cases yet.