Add Point expressions
PointBase expressions represent points in a 3D affine space.

Points are elements of AffineBase but not VectorBase. The main diffence is that
it is impossible to add two points, or scale or take the norm of a point.  I,
for one, welcome our new mathematically rigorous overlords.  Subtracting two
Points gives a Translation.

Add Subtract expression. Needed now instead of Sum<A, Minus<B>> because points
cannot be negated.

Add some internal template helpers to select the correct base of Sum and
Subtract regardless of argument order.