Improve error on mismatching Framed ctor
Previously trying to construct a Framed object with a mismatching expression
would result in a confusing "no matching constructor for..." error. Make
constructors accessible and add a static_assert for matching frames, nested in
a template so that the evaluated type of the rhs expression is shown in the
error message.

The new error message is 'static_assert failed "Mismatching frames"' with the
offending line 2 steps back in the trace.