Fix compile error in some jacobian calls
Fix #11

A "no matching constructor" error happened for a class of expressions
when the jacobian was requested via expr.jacobian(x), but not
expr.evalWithJacobians(x). The cause was incorrect template arguments
used to instantiate the jacobian evaluator.

Use correct and consistent evaluator type.  Add a test which catches the
issue.